requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
pymongo[srv]==3.6
scikit-learn
mlflow
pyaml
dagshub
fastapi
uvicorn
python-multipart

# enable the below once after the project is completed.
## -e .